refactor: migrate download genesis functionality back to cmd package
